+++ app/__init__.py
@@ -8,6 +8,8 @@ from .models import db, User
 from .api.user_routes import user_routes
 from .api.auth_routes import auth_routes
 from .api.business_routes import business_routes
+from .api.review_routes import review_routes
+from .api.business_image_routes import image_routes
 from .seeds import seed_commands
 from .config import Config
 
@@ -30,6 +32,8 @@ app.config.from_object(Config)
 app.register_blueprint(user_routes, url_prefix='/api/users')
 app.register_blueprint(auth_routes, url_prefix='/api/auth')
 app.register_blueprint(business_routes, url_prefix='/api/businesses')
+app.register_blueprint(review_routes, url_prefix='/api/reviews')
+app.register_blueprint(image_routes, url_prefix='/api/business-images')
 db.init_app(app)
 Migrate(app, db)
